frontend/fix: #409 added live stats dashboard in the side menu for driver and handling share app url's
