+++ Frontend/android-native/app/src/main/res/drawable/ic_graph_black.png
