+++ Frontend/ui-customer/src/Screens/CustomerUtils/ReferralScreen/View.purs
@@ -22,7 +22,7 @@ import Components.PrimaryEditText as PrimaryEditText
 import Effect (Effect)
 import Engineering.Helpers.Commons  as EHC
 import Font.Style as FontStyle
-import Helpers.Utils (adjustViewWithKeyboard, isPreviousVersion, getPreviousVersion)
+import Helpers.Utils (adjustViewWithKeyboard, getPreviousVersion)
 import Language.Strings (getString)
 import Language.Types (STR(..))
 import Prelude (Unit, bind, const, ($), (<<<), (<>), (==))
@@ -179,7 +179,7 @@ referenceView push state =
             [ imageView
                 [ height $ V 15
                 , width $ V 20
-                , imageUrl if state.isExpandReference then (if isPreviousVersion (getValueToLocalStore VERSION_NAME) (getPreviousVersion "") then "ic_chevron_up" else "ny_ic_chevron_up")  else (if isPreviousVersion (getValueToLocalStore VERSION_NAME) (getPreviousVersion "") then "ic_chevron_down" else "ny_ic_chevron_down")
+                , imageUrl if state.isExpandReference then (if EHC.isPreviousVersion (getValueToLocalStore VERSION_NAME) (getPreviousVersion "") then "ic_chevron_up" else "ny_ic_chevron_up")  else (if EHC.isPreviousVersion (getValueToLocalStore VERSION_NAME) (getPreviousVersion "") then "ic_chevron_down" else "ny_ic_chevron_down")
                 ]
             ]
         ]
