+++ Frontend/ui-driver/src/Resource/Localizable/Language/EN.purs
@@ -373,3 +373,4 @@ getEN script = case script of
   SILENT -> "Silent"
   GO_ONLINE -> "GO!"
   GO_ONLINE_PROMPT -> "You are currently offline.\nTo get ride requests, go online now!"
+  LIVE_DASHBOARD -> "Live Stats Dashboard"
