+++ Frontend/ui-driver/src/Resource/Localizable/Language/HI.purs
@@ -373,3 +373,4 @@ getHI script = case script of
   GO_SILENT -> "साइलेंट हो जाओ"
   GO_ONLINE -> "गो!"
   GO_ONLINE_PROMPT -> "आप वर्तमान में ऑफ़लाइन हैं।\nराइड अनुरोध प्राप्त करने के लिए, अभी ऑनलाइन हों!"
+  LIVE_DASHBOARD -> "लाइव आँकड़े डैशबोर्ड"
