+++ Frontend/ui-driver/src/Resource/Localizable/Language/KN.purs
@@ -373,3 +373,4 @@ getKN script = case script of
   GO_SILENT -> "ಸೈಲೆಂಟ್ ಸಕ್ರಿಯಗೊಳಿಸಿ"
   GO_ONLINE -> "ಗೋ!"
   GO_ONLINE_PROMPT -> "ನೀವು ಪ್ರಸ್ತುತ ಆಫ್‌ಲೈನ್‌ನಲ್ಲಿರುವಿರಿ. \n ಸವಾರಿ ವಿನಂತಿಗಳನ್ನು ಪಡೆಯಲು, ಈಗಲೇ ಆನ್‌ಲೈನ್‌ಗೆ ಹೋಗಿ!"
+  LIVE_DASHBOARD -> "ಲೈವ್ ಸ್ಥಿತಿ ಡ್ಯಾಶ್‌ಬೋರ್ಡ್"
