+++ Frontend/ui-driver/src/Resource/Localizable/Language/TA.purs
@@ -372,3 +372,4 @@ getTA script = case script of
   GO_SILENT -> "சைலண்ட் மோடுக்கு செல்லவும்"
   GO_ONLINE -> "போ!"
   GO_ONLINE_PROMPT -> "நீங்கள் தற்போது ஆஃப்லைனில் உள்ளீர்கள். \n சவாரிகளைப் பெற, இப்போதே ஆன்லைனில் செல்லுங்கள்!"
+  LIVE_DASHBOARD -> "நேரடி புள்ளிவிவரங்கள் டாஷ்போர்டு"
